Add providers list
