+++ app/src/main/java/com/github/wanderwise_inc/app/data/GoogleSignInLauncher.kt
@@ -1,5 +1,13 @@
 package com.github.wanderwise_inc.app.data
 
-interface GoogleSignInLauncher {
-  fun signIn()
+import android.content.Intent
+import androidx.activity.result.ActivityResultLauncher
+
+class DefaultGoogleSignInLauncher(
+    private val signInLauncher: ActivityResultLauncher<Intent>,
+    private val signInIntent: Intent
+) : GoogleSignInLauncher {
+  override fun signIn() {
+    signInLauncher.launch(signInIntent)
+  }
 }
