+++ app/src/main/java/com/github/wanderwise_inc/app/data/SignInLauncher.kt
@@ -1,0 +1,5 @@
+package com.github.wanderwise_inc.app.data
+
+interface GoogleSignInLauncher {
+  fun signIn()
+}
