+++ app/src/main/java/com/github/wanderwise_inc/app/data/SignInProviders.kt
@@ -1,0 +1,18 @@
+package com.github.wanderwise_inc.app.data
+
+import com.firebase.ui.auth.AuthUI
+
+object SignInProviders {
+    val providers by lazy {
+        listOf(
+            google
+        )
+    }
+
+    private val google by lazy {
+        AuthUI
+            .IdpConfig
+            .GoogleBuilder()
+            .build()
+    }
+}
